# [PoC] Add AVX2 version of get_sad