While playing around with rav1e I noticed it was spending a lot of time in the `get_sad` function, and vTune said further vectorization might help, so here's a rough hack to help rustc a little.

This should produce the same output as get_sad, and I think I've handled non-multiples of 8 sizes correctly, but this is definitely unsafe and worth double-checking!

As for results, rav1e now runs ~10% faster on my machine (encoding 24 frames at 1080p):

```
$ time ./target/release/rav1e -l 24 input.y4m -o output_short.ivf
```

|   | User | System | Real time |
| ------------- | ------------- | -- | -- |
| Before  | 128,33s | 0,37s | 2:08,73  |
| After  | 113,17s | 0,30s | 1:53,52 |