@tux3 2018-09-22
I've tried to factor out the common code in a `get_sad_generic`.

sse2 shouldn't be too hard to write, but rustc should already be able to autovectorize the iter/zip/map loop, it's just not the most efficient implementation.