@lu-zero 2018-09-25
I prepare [this](https://github.com/xiph/rav1e/pull/610/commits/028900dbebf13101731d9aa6ec6e84f418810512) to make easier to reason about get_sad.

Probably we should see if making specialized version helps. I'll try to come up with something the next days.