@KyleSiefring 2018-12-06
SSE assembly was added and the plan now is to do everything in assembly so I'm closing this for now.